requirements.txt
django==3.2.23
djangorestframework
requests
gunicorn
python-dotenv